fix: chunk overflow, atomic dispose, header pre-size, slowloris timeout
- HttpBodyParser: use (long)chunkSize + 2 to prevent int overflow DoS
- TcpNode/UdpNode/UdpDatagramLease/SocketIoEventArgsPool: replace bool
  _disposed with Interlocked.Exchange for thread-safe double-dispose
- HttpHeaderParser: pre-size List/Dict with capacity 16 to reduce GC
- HttpConnectionHandler: add RequestTimeout + CheckRequestTimeoutAsync
  to defend against slowloris attacks (default 30s)
- ConnectionRuntimeState: track RequestParsingStartedAtUtc per connection

+++ src/PicoNode.Http/Internal/HttpRequestParsing/HttpBodyParser.cs
@@ -148,7 +148,7 @@ internal static class HttpBodyParser
             }
 
             var remaining = reader.SliceFromPosition();
-            if (remaining.Length < chunkSize + 2)
+            if (remaining.Length < (long)chunkSize + 2)
             {
                 return HttpRequestParseResult.Incomplete(reader.BufferStart, expectsContinue);
             }
